fix: pdf generate
